Updating delete usecases

+++ src/modules/accounts/useCases/deleteCompany/DeleteCompanyUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 import { CompaniesRepository } from '@modules/accounts/infra/typeorm/repositories/CompaniesRepository';
+import { AppError } from '@shared/errors/AppError';
 
 @injectable()
 class DeleteCompanyUseCase {
@@ -10,6 +11,12 @@ class DeleteCompanyUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const company = await this.companiesRepository.findById(id);
+
+    if (!company) {
+      throw new AppError('Company does not exists!');
+    }
+
     await this.companiesRepository.deleteById(id);
   }
 }
