+++ src/modules/accounts/useCases/deleteDepartment/DeleteDepartmentUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 import { DepartmentsRepository } from '@modules/accounts/infra/typeorm/repositories/DepartmentsRepository';
+import { AppError } from '@shared/errors/AppError';
 
 @injectable()
 class DeleteDepartmentUseCase {
@@ -10,6 +11,12 @@ class DeleteDepartmentUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const department = await this.departmentsRepository.findById(id);
+
+    if (!department) {
+      throw new AppError('Department does not exists!');
+    }
+
     await this.departmentsRepository.deleteById(id);
   }
 }
