+++ src/modules/accounts/useCases/deleteGroup/DeleteGroupUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 import { GroupsRepository } from '@modules/accounts/infra/typeorm/repositories/GroupsRepository';
+import { AppError } from '@shared/errors/AppError';
 
 @injectable()
 class DeleteGroupUseCase {
@@ -10,6 +11,11 @@ class DeleteGroupUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const group = await this.groupsRepository.findById(id);
+
+    if (!group) {
+      throw new AppError('Group does not exists!');
+    }
     await this.groupsRepository.deleteById(id);
   }
 }
