+++ src/modules/accounts/useCases/deletePosition/DeletePositionUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 import { PositionsRepository } from '@modules/accounts/infra/typeorm/repositories/PositionsRepository';
+import { AppError } from '@shared/errors/AppError';
 
 @injectable()
 class DeletePositionUseCase {
@@ -10,6 +11,12 @@ class DeletePositionUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const position = await this.positionsRepository.findById(id);
+
+    if (!position) {
+      throw new AppError('Position does not exists!');
+    }
+
     await this.positionsRepository.deleteById(id);
   }
 }
