+++ src/modules/accounts/useCases/deleteUser/DeleteUserUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
+import { AppError } from '@shared/errors/AppError';
 
 @injectable()
 class DeleteUserUseCase {
@@ -10,6 +11,12 @@ class DeleteUserUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const user = await this.usersRepository.findById(id);
+
+    if (!user) {
+      throw new AppError('User does not exists!');
+    }
+
     await this.usersRepository.deleteById(id);
   }
 }
